# internal/models/resources.go
package models

type UserCreate struct {
	Email    string `json:"email" binding:"required"`
	Password string `json:"password" binding:"required"`
}

type UserUpdate struct {
	Name string `json:"name"`
}